address mathieu comments 2.0
